[time] new nmea_gettime function
This function is a wrapper of clock_gettime and use the time of the GPS
informations.
By default this function open the /dev/ttyS0 device to recieve the data.
The path may be changed with the environment variable NMEA_TTYGPS.

With GCC, the dynamic loader wrap clock_gettime and a fallback is
available to return to the generic function if there isn't more than 2 satellites in use.